+++ slam3d/particlefilter/particleFilter.h
@@ -71,10 +71,10 @@ extern "C" {
 	} particleFilter_t;
 
 	void particleFilter_init(particleFilter_t* pf);
-	void particleFilter_depositVio(particleFilter_t* pf, float t, float x, float y, float z, float dist);
+	void particleFilter_depositVio(particleFilter_t* pf, double t, float x, float y, float z, float dist);
     void particleFilter_depositUwb(particleFilter_t* pf, bcn_t* bcn, float range, float stdRange);
-    void particleFilter_getTagLoc(const particleFilter_t* pf, float* t, float* x, float* y, float* z, float* theta);
-    void particleFilter_getBcnLoc(const particleFilter_t* pf, const bcn_t* bcn, float* t, float* x, float* y, float* z);
+    void particleFilter_getTagLoc(const particleFilter_t* pf, double* t, float* x, float* y, float* z, float* theta);
+    void particleFilter_getBcnLoc(const particleFilter_t* pf, const bcn_t* bcn, double* t, float* x, float* y, float* z);
 
 #ifdef __cplusplus
 } // extern "C"
